refactoring for unit tests
