+++ src/main/resources/application-prod.yml
@@ -21,3 +21,5 @@ spring:
     problemdetails:
       enabled: false
 
+server:
+  port : 8082
