+++ src/main/resources/application.yml
@@ -1,7 +1,6 @@
 spring:
   profiles:
     active: dev
-server:
-  port : 8081
+
